requirements.txt
streamlit
torch
matplotlib
numpy
pillow
requests
